+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Allow syncing `null` document data.
+
 ## [0.2.0-beta.1] - 2024-05-14
 
 ### ⚠️ Breaking Changes